Fix CLI usage
