+++ package.json
@@ -11,10 +11,10 @@
     "url": "https://github.com/parzh/create-package-javascript/issues",
     "email": "parzhitsky@gmail.com"
   },
-  "bin": "src/index.js",
+  "bin": "src/index",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --unhandled-rejections=strict --require ts-node/register src",
+    "start": ".\\src\\index",
     "postpublish": "git push --follow-tags"
   },
   "devDependencies": {
